Update QScheme data type to 16-bit unsigned integer.
This commit changes the QScheme data type to a 16-bit unsigned integer.
This is to have consistency with the BCQTensor in reading and saving quantization information, which eases calculating the offset.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/tensor/quantizer.h
@@ -29,7 +29,7 @@ class Tensor;
  * updated. If you would like to use a different quantization technique, please
  * select a custom quantizer scheme.
  */
-enum class QScheme : uint8_t {
+enum class QScheme : uint16_t {
   /** predefined quantizer */
   PER_TENSOR_AFFINE = 0x00,
   PER_CHANNEL_AFFINE = 0x01,
